criacao de calculadora

+++ quebecbottcamp/Primeiro.Programa/src/br/com/dio/calculadora/Calculadora.java
@@ -17,7 +17,7 @@ public class Calculadora {
 		int soma = soma(a,b);
 	    int subtracao = subtracao(a,b);
 	    int multiplicacao = multiplicacao(a,b);
-	    int divisao = divisao(a,b);
+	    float divisao = divisao(a,b);
 	    
 	    System.out.println("soma" + soma);
 	    System.out.println("sub" + subtracao);
@@ -33,7 +33,7 @@ public class Calculadora {
 			return a - b;
 	 }
 		
-		public static int divisao(int a, int b) {
+		public static float divisao(float a, float b) {
 			return a / b;
 	 }
